Сделал рефакторинг, и добавил отрисовку попапа при открытии
