+++ src/main.js
@@ -1,9 +1,9 @@
+import FilmsModel from './model/films-model.js';
 import HeaderProfileView from './view/header-profile-view.js';
 import MainNavigationView from './view/main-navigation-view.js';
-import FilmsPresenter from './presenter/films-presenter.js';
 import FooterStatisticsView from './view/footer-statistics-view.js';
+import FilmsPresenter from './presenter/films-presenter.js';
 import {render} from './render.js';
-import FilmsModel from './model/films-model.js';
 
 
 const siteHeaderElement = document.querySelector('.header');
@@ -20,3 +20,9 @@ render(new MainNavigationView(), siteMainElement);
 render(new FooterStatisticsView(), siteFooterElement);
 
 filmsPresenter.init(siteMainElement, filmsModel);
+
+//  Стоит ли во view, film передавать #film?
+//  Как проверить копится ли елемент в оперативной памяти
+//  Что значит верхнеуровневая копия
+//  Не могу проверить копятся ли обработчики событий
+//  Как удалить обработчики
