+++ src/view/film-card-view.js
@@ -34,18 +34,19 @@ const createFilmCardTemplate = (film) => {
 
 export default class FilmCardView {
   #element = null;
+  #film = null;
 
   constructor(film) {
-    this.film = film;
+    this.#film = film;
   }
 
-  getTemplate() {
-    return createFilmCardTemplate(this.film);
+  get template() {
+    return createFilmCardTemplate(this.#film);
   }
 
   getElement() {
     if(!this.#element) {
-      this.#element = createElement(this.getTemplate());
+      this.#element = createElement(this.template);
     }
 
     return this.#element;
