--- src/view/film-details-bottom-container-view.js
@@ -1,98 +1,0 @@
-import { createElement } from '../render.js';
-
-const createFilmDetailsBottomContainerTemplate = (film, commentsArray) => {
-  const { comments } = film;
-
-  const addComments = () => {
-    let commentsList = '';
-
-    commentsArray.forEach((data) => {
-      const { author, comment, date, emotion } = data;
-
-      commentsList += `<li class="film-details__comment">
-            <span class="film-details__comment-emoji">
-              <img src="./images/emoji/${emotion}.png" width="55" height="55" alt="emoji-smile">
-            </span>
-            <div>
-              <p class="film-details__comment-text">${comment}</p>
-              <p class="film-details__comment-info">
-                <span class="film-details__comment-author">${author}</span>
-                <span class="film-details__comment-day">${date}</span>
-                <button class="film-details__comment-delete">Delete</button>
-              </p>
-           </div>
-         </li>`;
-    });
-
-    return commentsList;
-  };
-
-
-  return (
-    `<div class="film-details__bottom-container">
-      <section class="film-details__comments-wrap">
-        <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${comments.length}</span></h3>
-
-        <ul class="film-details__comments-list">
-        ${addComments()}
-        </ul>
-
-        <div class="film-details__new-comment">
-          <div class="film-details__add-emoji-label"></div>
-
-          <label class="film-details__comment-label">
-            <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
-          </label>
-
-          <div class="film-details__emoji-list">
-            <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-smile" value="smile">
-            <label class="film-details__emoji-label" for="emoji-smile">
-              <img src="./images/emoji/smile.png" width="30" height="30" alt="emoji">
-            </label>
-
-            <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-sleeping" value="sleeping">
-            <label class="film-details__emoji-label" for="emoji-sleeping">
-              <img src="./images/emoji/sleeping.png" width="30" height="30" alt="emoji">
-            </label>
-
-            <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-puke" value="puke">
-            <label class="film-details__emoji-label" for="emoji-puke">
-              <img src="./images/emoji/puke.png" width="30" height="30" alt="emoji">
-            </label>
-
-            <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-angry" value="angry">
-            <label class="film-details__emoji-label" for="emoji-angry">
-              <img src="./images/emoji/angry.png" width="30" height="30" alt="emoji">
-            </label>
-          </div>
-        </div>
-      </section>
-    </div>`
-  );
-};
-
-
-export default class FilmDetailsBottomContainerView {
-  #element = null;
-
-  constructor(film, commentsArray) {
-    this.film = film;
-    this.commentsArray = commentsArray;
-  }
-
-  getTemplate() {
-    return createFilmDetailsBottomContainerTemplate(this.film, this.commentsArray);
-  }
-
-  getElement() {
-    if(!this.#element) {
-      this.#element = createElement(this.getTemplate());
-    }
-
-    return this.#element;
-  }
-
-  removeElement() {
-    this.#element = null;
-  }
-}
