--- src/view/film-details-top-container-view.js
@@ -1,111 +1,0 @@
-import { createElement } from '../render.js';
-
-const createFilmDetailsTopContainerTemplate = (film) => {
-  const { title, totalRating, poster, release, description, genre, ageRating, alternativeTitle, director, writers, actors } = film.filmInfo;
-  const { watchList, alreadyWatched, favorite } = film.userDetails;
-
-  const addControlsActiveClassName = (control) => (control) ? 'film-details__control-button--active' : '';
-
-  const addGenre = () => {
-    let template = '';
-
-    for (let i = 0; i < genre.length; i++) {
-      template += `<span class="film-details__genre">${genre[i]}</span>`;
-    }
-
-    return template;
-  };
-
-  return (
-    `<div class="film-details__top-container">
-      <div class="film-details__close">
-        <button class="film-details__close-btn" type="button">close</button>
-      </div>
-      <div class="film-details__info-wrap">
-        <div class="film-details__poster">
-          <img class="film-details__poster-img" src="${poster}">
-
-          <p class="film-details__age">${ageRating}+</p>
-        </div>
-
-        <div class="film-details__info">
-          <div class="film-details__info-head">
-            <div class="film-details__title-wrap">
-              <h3 class="film-details__title">${title}</h3>
-              <p class="film-details__title-original">Original: ${alternativeTitle}</p>
-            </div>
-
-            <div class="film-details__rating">
-              <p class="film-details__total-rating">${totalRating}</p>
-            </div>
-          </div>
-
-          <table class="film-details__table">
-            <tr class="film-details__row">
-              <td class="film-details__term">Director</td>
-              <td class="film-details__cell">${director}</td>
-            </tr>
-            <tr class="film-details__row">
-              <td class="film-details__term">Writers</td>
-              <td class="film-details__cell">${writers}</td>
-            </tr>
-            <tr class="film-details__row">
-              <td class="film-details__term">Actors</td>
-              <td class="film-details__cell">${actors.join(', ')}</td>
-            </tr>
-            <tr class="film-details__row">
-              <td class="film-details__term">Release Date</td>
-              <td class="film-details__cell">30 March 1945</td>
-            </tr>
-            <tr class="film-details__row">
-              <td class="film-details__term">Runtime</td>
-              <td class="film-details__cell">1h 18m</td>
-            </tr>
-            <tr class="film-details__row">
-              <td class="film-details__term">Country</td>
-              <td class="film-details__cell">${release.releaseCountry}</td>
-            </tr>
-            <tr class="film-details__row">
-              <td class="film-details__term">Genres</td>
-              <td class="film-details__cell">
-                ${addGenre()}
-              </td>
-            </tr>
-          </table>
-
-          <p class="film-details__film-description">${description}</p>
-        </div>
-      </div>
-
-      <section class="film-details__controls">
-        <button type="button" class="film-details__control-button ${addControlsActiveClassName(watchList)} film-details__control-button--watchlist" id="watchlist" name="watchlist">Add to watchlist</button>
-        <button type="button" class="film-details__control-button ${addControlsActiveClassName(alreadyWatched)} film-details__control-button--watched" id="watched" name="watched">Already watched</button>
-        <button type="button" class="film-details__control-button ${addControlsActiveClassName(favorite)} film-details__control-button--favorite" id="favorite" name="favorite">Add to favorites</button>
-      </section>
-    </div>`
-  );
-};
-
-export default class FilmDetailsTopContainerView {
-  #element = null;
-
-  constructor(film) {
-    this.film = film;
-  }
-
-  getTemplate() {
-    return createFilmDetailsTopContainerTemplate(this.film);
-  }
-
-  getElement() {
-    if(!this.#element) {
-      this.#element = createElement(this.getTemplate());
-    }
-
-    return this.#element;
-  }
-
-  removeElement() {
-    this.#element = null;
-  }
-}
