--- src/view/film-details-view.js
@@ -1,23 +1,0 @@
-import { createElement } from '../render.js';
-
-const createFilmDetailsTemplate = () => '<section class="film-details"></section>';
-
-export default class FilmDetailsView {
-  #element = null;
-
-  getTemplate() {
-    return createFilmDetailsTemplate();
-  }
-
-  getElement() {
-    if(!this.#element) {
-      this.#element = createElement(this.getTemplate());
-    }
-
-    return this.#element;
-  }
-
-  removeElement() {
-    this.#element = null;
-  }
-}
