+++ src/view/show-more-button-view.js
@@ -5,13 +5,13 @@ const createShowMoreButtonTemplate = () => '<button class="films-list__show-more
 export default class ShowMoreButtonView {
   #element = null;
 
-  getTemplate() {
+  get template() {
     return createShowMoreButtonTemplate();
   }
 
   getElement() {
     if(!this.#element) {
-      this.#element = createElement(this.getTemplate());
+      this.#element = createElement(this.template);
     }
 
     return this.#element;
