+++ src/view/sort-view.js
@@ -11,13 +11,13 @@ const createSortTemplate = () => (
 export default class SortView {
   #element = null;
 
-  getTemplate() {
+  get template() {
     return createSortTemplate();
   }
 
   getElement() {
     if(!this.#element) {
-      this.#element = createElement(this.getTemplate());
+      this.#element = createElement(this.template);
     }
 
     return this.#element;
